Updated for use with Movable Type 5.2.x
Rewrote the PromoteThis for use in Movable Type 5.2.3 and later.  This
was done to facilitate upgrades of older Movable Type-based sites, and
in order to avoid continued use of Movable Type 4.  Movable Type 4.38
is going End-Of-Life at the end of 2013.

The license was changed from Artistic License (Perl) to MIT License by
After6 Services as the copyright holder and developer of the Movable
Type 5 version of this plugin.